requirements.txt
huggingface_hub
fastapi[standard]
google-genai
groq
xai-sdk
Pillow
uvicorn 
jinja2 
python-multipart






